Add support for new talent specializations
This new version works a bit differently than the old one in that the
same spell can be bound to multiple talent sets and it should work
correctly. Basically if a talent bind set is selected then the bind will
only be available when that spec is active, but it is no longer an ONLY
bind set.