Attempt to resolve the issue with "stuck" bindings
There are some situations in which the mouseover unit can go away, and
we're currently using this to detect if a frame might be obscuring the
current one or if the current frame has been hidden.

I have made this check smarter and it should be better at determining
when to clear the bindings that have been set. If you want to view more
information about when bindings are being clear (or not) you can run the
following in-game:

/run Clique.settings.debugUnitIssue = true
/reload

You can remove these messages later with the reverse:

/run Clique.settings.debugUnitIssue = false
/reload