Fire bindings changed on player entering world
Some spec-based bindings are being lost in between instances, this
should help to mitigate that. It's difficult to reproduce without the
right character/instance which I'm not able to reproduce but this should
fix some of those issues.

It will cause Clique to do a slight bit more work when zoning between
instances, but I don't think there are any crucial points where that
happens.